LayoutTests/fast/images/text-recognition/image-overlay-size-change-expected.txt

PASS width is 200
PASS height is 200
PASS width became 50
PASS height is 50
PASS width became 200
PASS height is 200
PASS successfullyParsed is true

TEST COMPLETE
